+++ src/docs/constants/false.js
@@ -1,0 +1,12 @@
+math.docs['false'] = {
+    'name': 'false',
+    'category': 'Constants',
+    'syntax': [
+        'false'
+    ],
+    'description': 'Boolean value false',
+    'examples': [
+        'false'
+    ],
+    'seealso': ['true']
+};