+++ src/docs/constants/true.js
@@ -1,0 +1,12 @@
+math.docs['true'] = {
+    'name': 'true',
+    'category': 'Constants',
+    'syntax': [
+        'true'
+    ],
+    'description': 'Boolean value true',
+    'examples': [
+        'true'
+    ],
+    'seealso': ['false']
+};