+++ src/docs/function/construction/boolean.js
@@ -1,0 +1,20 @@
+math.docs['boolean'] = {
+    'name': 'boolean',
+    'category': 'Type',
+    'syntax': [
+        'x',
+        'boolean(x)'
+    ],
+    'description':
+        'Convert a string or number into a boolean.',
+    'examples': [
+        'boolean(0)',
+        'boolean(1)',
+        'boolean(3)',
+        'boolean("true")',
+        'boolean("false")'
+    ],
+    'seealso': [
+        'complex', 'matrix', 'range', 'string', 'unit'
+    ]
+};