+++ src/docs/function/utils/help.js
@@ -2,12 +2,13 @@ math.docs.help = {
     'name': 'help',
     'category': 'Utils',
     'syntax': [
-        'help(object)'
+        'help(object)',
+        'help(string)'
     ],
     'description': 'Display documentation on a function or data type.',
     'examples': [
-        'help("sqrt");',
-        'help("complex");'
+        'help(sqrt)',
+        'help("complex")'
     ],
     'seealso': []
 };